+++ src/main/java/com/raoulvdberge/refinedstorage/RSConfig.java
@@ -72,6 +72,7 @@ public class RSConfig {
 
     //region Wireless Grid
     public boolean wirelessGridUsesEnergy;
+    public int wirelessGridCapacity;
     public int wirelessGridOpenUsage;
     public int wirelessGridExtractUsage;
     public int wirelessGridInsertUsage;
@@ -79,6 +80,7 @@ public class RSConfig {
 
     //region Portable Grid
     public boolean portableGridUsesEnergy;
+    public int portableGridCapacity;
     public int portableGridOpenUsage;
     public int portableGridExtractUsage;
     public int portableGridInsertUsage;
@@ -86,6 +88,7 @@ public class RSConfig {
 
     //region Wireless Fluid Grid
     public boolean wirelessFluidGridUsesEnergy;
+    public int wirelessFluidGridCapacity;
     public int wirelessFluidGridOpenUsage;
     public int wirelessFluidGridExtractUsage;
     public int wirelessFluidGridInsertUsage;
@@ -93,6 +96,7 @@ public class RSConfig {
 
     //region Wireless Crafting Monitor
     public boolean wirelessCraftingMonitorUsesEnergy;
+    public int wirelessCraftingMonitorCapacity;
     public int wirelessCraftingMonitorOpenUsage;
     public int wirelessCraftingMonitorCancelUsage;
     public int wirelessCraftingMonitorCancelAllUsage;
@@ -207,6 +211,7 @@ public class RSConfig {
 
         //region Wireless Grid
         wirelessGridUsesEnergy = config.getBoolean("usesEnergy", WIRELESS_GRID, true, "Whether the Wireless Grid uses energy");
+        wirelessGridCapacity = config.getInt("capacity", WIRELESS_GRID, 3200, 0, Integer.MAX_VALUE, "The energy capacity of the Wireless Grid");
         wirelessGridOpenUsage = config.getInt("open", WIRELESS_GRID, 30, 0, Integer.MAX_VALUE, "The energy used by the Wireless Grid to open");
         wirelessGridInsertUsage = config.getInt("insert", WIRELESS_GRID, 3, 0, Integer.MAX_VALUE, "The energy used by the Wireless Grid to insert items");
         wirelessGridExtractUsage = config.getInt("extract", WIRELESS_GRID, 3, 0, Integer.MAX_VALUE, "The energy used by the Wireless Grid to extract items");
@@ -214,13 +219,15 @@ public class RSConfig {
 
         //region Portable Grid
         portableGridUsesEnergy = config.getBoolean("usesEnergy", PORTABLE_GRID, true, "Whether the Portable Grid uses energy");
+        portableGridCapacity = config.getInt("capacity", PORTABLE_GRID, 3200, 0, Integer.MAX_VALUE, "The energy capacity of the Portable Grid");
         portableGridOpenUsage = config.getInt("open", PORTABLE_GRID, 30, 0, Integer.MAX_VALUE, "The energy used by the Portable Grid to open");
         portableGridInsertUsage = config.getInt("insert", PORTABLE_GRID, 3, 0, Integer.MAX_VALUE, "The energy used by the Portable Grid to insert items");
         portableGridExtractUsage = config.getInt("extract", PORTABLE_GRID, 3, 0, Integer.MAX_VALUE, "The energy used by the Portable Grid to extract items");
         //endregion
 
-        //region Wireless Grid
+        //region Wireless Fluid Grid
         wirelessFluidGridUsesEnergy = config.getBoolean("usesEnergy", WIRELESS_FLUID_GRID, true, "Whether the Fluid Wireless Grid uses energy");
+        wirelessFluidGridCapacity = config.getInt("capacity", WIRELESS_FLUID_GRID, 3200, 0, Integer.MAX_VALUE, "The energy capacity of the Wireless Fluid Grid");
         wirelessFluidGridOpenUsage = config.getInt("open", WIRELESS_FLUID_GRID, 30, 0, Integer.MAX_VALUE, "The energy used by the Fluid Wireless Grid to open");
         wirelessFluidGridInsertUsage = config.getInt("insert", WIRELESS_FLUID_GRID, 3, 0, Integer.MAX_VALUE, "The energy used by the Wireless Fluid Grid to insert items");
         wirelessFluidGridExtractUsage = config.getInt("extract", WIRELESS_FLUID_GRID, 3, 0, Integer.MAX_VALUE, "The energy used by the Wireless Fluid Grid to extract items");
@@ -228,6 +235,7 @@ public class RSConfig {
 
         //region Wireless Crafting Monitor
         wirelessCraftingMonitorUsesEnergy = config.getBoolean("usesEnergy", WIRELESS_CRAFTING_MONITOR, true, "Whether the Wireless Crafting Monitor uses energy");
+        wirelessCraftingMonitorCapacity = config.getInt("capacity", WIRELESS_CRAFTING_MONITOR, 3200, 0, Integer.MAX_VALUE, "The energy capacity of the Wireless Crafting Monitor");
         wirelessCraftingMonitorOpenUsage = config.getInt("open", WIRELESS_CRAFTING_MONITOR, 35, 0, Integer.MAX_VALUE, "The energy used by the Wireless Crafting Monitor to open");
         wirelessCraftingMonitorCancelUsage = config.getInt("cancel", WIRELESS_CRAFTING_MONITOR, 4, 0, Integer.MAX_VALUE, "The energy used by the Wireless Crafting Monitor to cancel a task");
         wirelessCraftingMonitorCancelAllUsage = config.getInt("cancelAll", WIRELESS_CRAFTING_MONITOR, 5, 0, Integer.MAX_VALUE, "The energy used by the Wireless Crafting Monitor to cancel all tasks");